+++ Controller/Adminhtml/Webhooks/RegisterApplepay.php
@@ -7,13 +7,14 @@ use Magento\Framework\Controller\Result\Redirect;
 use Magento\Backend\App\Action;
 use Magento\Framework\HTTP\Client\Curl;
 use Magento\Framework\App\Config\Storage\WriterInterface;
-use Magento\Store\Model\StoreManagerInterface;
 use Magento\Config\Model\ResourceModel\Config\Data\CollectionFactory as configCollection;
+use UnzerSDK\Services\EnvironmentService;
 use UnzerSDK\Unzer;
+use UnzerSDK\Services\HttpService;
 use Magento\Backend\Model\Auth\Session;
 
 /**
- * Controller for registering webhooks via the backend
+ * Controller for registering Certificates via the backend
  *
  * @link  https://docs.unzer.com/
  *
@@ -22,40 +23,37 @@ use Magento\Backend\Model\Auth\Session;
 class RegisterApplepay extends Action
 {
 
+    private const URL_PART_STAGING_ENVIRONMENT = 'stg';
+    private const URL_PART_DEVELOPMENT_ENVIRONMENT = 'dev';
+    private const URL_PART_SANDBOX_ENVIRONMENT = 'sbx';
+
+
     private Curl $curl;
     private ScopeConfigInterface $_scopeConfig;
     private WriterInterface $_configWriter;
-    private StoreManagerInterface $_storeManager;
     private configCollection $configCollection;
+    private HttpService $httpService;
 
 
     public function __construct(
-        Action\Context        $context,
-        ScopeConfigInterface  $scopeConfig,
-        Curl                  $curl,
-        WriterInterface       $configWriter,
-        StoreManagerInterface $storeManager,
-        ConfigCollection      $configCollection,
-        Session               $session
+        Action\Context       $context,
+        ScopeConfigInterface $scopeConfig,
+        Curl                 $curl,
+        WriterInterface      $configWriter,
+        ConfigCollection     $configCollection,
+        HttpService          $httpService,
+        Session              $session
     )
     {
         parent::__construct($context);
         $this->_scopeConfig = $scopeConfig;
         $this->curl = $curl;
         $this->_configWriter = $configWriter;
-        $this->_storeManager = $storeManager;
         $this->configCollection = $configCollection;
+        $this->httpService = $httpService;
         $this->_session = $session;
     }
 
-    public function getApiUrl($logging, $url)
-    {
-        $envPrefix = 'sbx-';
-        if (!$logging) {
-            $envPrefix = '';
-        }
-        return "https://" . $envPrefix . Unzer::BASE_URL . "/" . $url;
-    }
 
     /**
      * @inheritDoc
@@ -64,12 +62,12 @@ class RegisterApplepay extends Action
     {
         $this->getRequest()->getParams();
         $mode = $this->getRequest()->getParam('switch');
-        $storeId = (int) $this->getRequest()->getParam('store', 0);
+        $storeId = (int)$this->getRequest()->getParam('store', 0);
         $publicKey = $this->_scopeConfig->getValue('payment/unzer/public_key', 'store', $storeId);
-        $unzerPrivateKey = base64_encode($this->_scopeConfig->getValue('payment/unzer/private_key', 'store', $storeId) . ":");
+        $privateKey = $this->_scopeConfig->getValue('payment/unzer/private_key', 'store', $storeId);
+        $unzerPrivateKey = base64_encode($privateKey . ":");
         $this->curl->addHeader("Content-Type", "application/json");
         $this->curl->addHeader("Authorization", "Basic " . $unzerPrivateKey);
-        $result = '';
 
         $sslCertPath = '../app/etc/upload/applepay_csr/' . $this->_scopeConfig->getValue('payment/unzer/applepay/csr_certificate_upload', 'store', $storeId);
         $sslCert = file_get_contents($sslCertPath);
@@ -84,11 +82,12 @@ class RegisterApplepay extends Action
         $sslKey = str_replace('-----END PRIVATE KEY-----', '', $sslKey);
         $sslKey = str_replace("\n", "", $sslKey);
 
+
         switch ($mode) {
 
             case 'registerPrivateKey':
                 $certificate = $sslKey;
-                $url = $this->getApiUrl($this->_scopeConfig->getValue('payment/unzer/logging', 'store', $storeId), 'v1/keypair/applepay/privatekeys');
+                $url = $this->getApiUrl($privateKey, 'v1/keypair/applepay/privatekeys');
 
                 $params = [
                     'format' => 'PEM',
@@ -99,7 +98,7 @@ class RegisterApplepay extends Action
                 // post method
                 $this->curl->post($url, json_encode($params));
 
-                // output of curl requestt
+                // output of curl request
                 $result = $this->curl->getBody();
 
                 $result = (array)json_decode($result);
@@ -110,16 +109,15 @@ class RegisterApplepay extends Action
                     $this->messageManager->addErrorMessage(__($errorMessage));
                 } else {
                     $this->_configWriter->save('payment/unzer_applepay/csr_private_key_response', $result['id'], $scope = ScopeConfigInterface::SCOPE_TYPE_DEFAULT, $scopeId = 0);
-                    $this->messageManager->addSuccessMessage(__('Successfully registered for public key: '.$publicKey));
+                    $this->messageManager->addSuccessMessage(__('Successfully registered for public key: ' . $publicKey));
                 }
                 break;
 
             case 'registerCertificate':
 
                 $certificate = $sslCert;
-                //$privateKey = $this->_scopeConfig->getValue('payment/unzer_applepay/csr_private_key_response');
                 $privateKey = $this->getRegistrationResponseValue("payment/unzer_applepay/csr_private_key_response");
-                $url = $this->getApiUrl($this->_scopeConfig->getValue('payment/unzer/logging', 'store', $storeId), 'v1/keypair/applepay/certificates');
+                $url = $this->getApiUrl($privateKey, 'v1/keypair/applepay/certificates');
 
                 $params = [
                     'format' => 'PEM',
@@ -136,20 +134,13 @@ class RegisterApplepay extends Action
 
                 $result = (array)json_decode($result);
 
-                if (array_key_exists('errors', $result)) {
-                    $errors = $result['errors'][0];
-                    $errorMessage = $errors->code . " - " . $errors->merchantMessage;
-                    $this->messageManager->addErrorMessage(__($errorMessage));
-                } else {
-                    $this->_configWriter->save('payment/unzer_applepay/csr_certificate_response', $result['id'], $scope = ScopeConfigInterface::SCOPE_TYPE_DEFAULT, $scopeId = 0);
-                    $this->messageManager->addSuccessMessage(__('Successfully registered for public key: '.$publicKey));
-                }
+                $this->handleResult($result,$publicKey);
                 break;
 
             case 'activate':
                 $certificateId = $this->getRegistrationResponseValue("payment/unzer_applepay/csr_certificate_response");
 
-                $url = $this->getApiUrl($this->_scopeConfig->getValue('payment/unzer/logging', 'store', $storeId), 'v1/keypair/applepay/certificates/' . $certificateId . '/activate');
+                $url = $this->getApiUrl($privateKey, 'v1/keypair/applepay/certificates/' . $certificateId . '/activate');
 
                 // post method
                 $this->curl->post($url, []);
@@ -159,34 +150,47 @@ class RegisterApplepay extends Action
 
                 $result = (array)json_decode($result);
 
-                if (array_key_exists('errors', $result)) {
-                    $errors = $result['errors'][0];
-                    $errorMessage = $errors->code . " - " . $errors->merchantMessage;
-                    $this->messageManager->addErrorMessage(__($errorMessage));
-                } else {
-                    $this->_configWriter->save('payment/unzer_applepay/csr_certificate_response', $result['id'], $scope = ScopeConfigInterface::SCOPE_TYPE_DEFAULT, $scopeId = 0);
-                    $this->messageManager->addSuccessMessage(__('Successfully registered for public key: '.$publicKey));
-                }
+                $this->handleResult($result,$publicKey);
                 break;
         }
         $redirect = $this->resultRedirectFactory->create();
-        $redirect->setUrl($this->getRequest()->getHeader('Referer').$this->getAnchorLink());
+        $redirect->setUrl($this->getRequest()->getHeader('Referer') . $this->getAnchorLink());
         return $redirect;
     }
 
     /**
+     * @param $result
+     * @param $publicKey
+     * @return void
+     */
+    public function handleResult($result,$publicKey){
+        if (array_key_exists('errors', $result)) {
+            $errors = $result['errors'][0];
+            $errorMessage = $errors->code . " - " . $errors->merchantMessage;
+            $this->messageManager->addErrorMessage(__($errorMessage));
+        } else {
+            $this->_configWriter->save('payment/unzer_applepay/csr_certificate_response', $result['id'], $scope = ScopeConfigInterface::SCOPE_TYPE_DEFAULT, $scopeId = 0);
+            $this->messageManager->addSuccessMessage(__('Successfully registered for public key: ' . $publicKey));
+        }
+    }
+    /**
      * @return string
      */
-    public function getAnchorLink(){
+    public function getAnchorLink(): string
+    {
         $anchor = "#payment_us_unzer_applepay-link";
-        $configState = $this->_session->getUser()->getExtra();
-        $configStateKeys = array_keys($configState["configState"]);
+        $configState = (array) $this->_session->getUser()->getExtra();
+        $configStateKeys = '';
+        if (!empty($configState) && array_key_exists('configState', $configState)) {
+            $configStateKeys = array_keys($configState['configState']);
+        }
         $dataForAnchorGeneration = preg_grep('/.*unzer.*/', $configStateKeys);
-        if(!empty($dataForAnchorGeneration)){
-            $anchor = "#".$dataForAnchorGeneration[0]."_applepay-link";
+        if (!empty($dataForAnchorGeneration)) {
+            $anchor = "#" . $dataForAnchorGeneration[0] . "_applepay-link";
         }
         return $anchor;
     }
+
     /**
      * @return mixed|string
      */
@@ -200,5 +204,51 @@ class RegisterApplepay extends Action
             return "";
         }
 
+    }
+
+    /**
+     * @param $privateKey
+     * @param $url
+     * @return string
+     */
+    public function getApiUrl($privateKey, $url): string
+    {
+        $envPrefix = $this->getEnvironmentPrefix($privateKey);
+        return "https://" . $envPrefix . Unzer::BASE_URL . "/" . $url;
+    }
+
+    /**
+     * @param $privateKey
+     * @return string
+     */
+    function getEnvironmentPrefix($privateKey): string
+    {
+        // Production Environment uses no prefix.
+        if ($this->isProductionKey($privateKey)) {
+            return '';
+        }
+
+        switch ($this->httpService->getEnvironmentService()->getPapiEnvironment()) {
+            case EnvironmentService::ENV_VAR_VALUE_STAGING_ENVIRONMENT:
+                $envPrefix = self::URL_PART_STAGING_ENVIRONMENT;
+                break;
+            case EnvironmentService::ENV_VAR_VALUE_DEVELOPMENT_ENVIRONMENT:
+                $envPrefix = self::URL_PART_DEVELOPMENT_ENVIRONMENT;
+                break;
+            default:
+                $envPrefix = self::URL_PART_SANDBOX_ENVIRONMENT;
+        }
+        return $envPrefix . '-';
+    }
+
+    /** Determine whether key is for production environment.
+     *
+     * @param string $privateKey
+     *
+     * @return bool
+     */
+    private function isProductionKey(string $privateKey): bool
+    {
+        return strpos($privateKey, 'p') === 0;
     }
 }